requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
slack_bolt>=1.18.0
playwright>=1.40.0
Pillow>=10.0.0

